REFACTOR

+++ src/main/java/org/example/expert/domain/manager/controller/ManagerController.java
@@ -4,6 +4,7 @@ import io.jsonwebtoken.Claims;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.expert.config.JwtUtil;
+import org.example.expert.domain.auth.exception.AuthException;
 import org.example.expert.domain.common.annotation.Auth;
 import org.example.expert.domain.common.dto.AuthUser;
 import org.example.expert.domain.manager.dto.request.ManagerSaveRequest;
@@ -17,12 +18,13 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/todos/{todoId}/managers")
 public class ManagerController {
 
     private final ManagerService managerService;
     private final JwtUtil jwtUtil;
 
-    @PostMapping("/todos/{todoId}/managers")
+    @PostMapping
     public ResponseEntity<ManagerSaveResponse> saveManager(
             @Auth AuthUser authUser,
             @PathVariable long todoId,
@@ -31,18 +33,26 @@ public class ManagerController {
         return ResponseEntity.ok(managerService.saveManager(authUser, todoId, managerSaveRequest));
     }
 
-    @GetMapping("/todos/{todoId}/managers")
+    @GetMapping
     public ResponseEntity<List<ManagerResponse>> getMembers(@PathVariable long todoId) {
         return ResponseEntity.ok(managerService.getManagers(todoId));
     }
 
-    @DeleteMapping("/todos/{todoId}/managers/{managerId}")
+    @DeleteMapping("/{managerId}")
     public void deleteManager(
             @RequestHeader("Authorization") String bearerToken,
             @PathVariable long todoId,
             @PathVariable long managerId
     ) {
+        if (bearerToken == null || !bearerToken.startsWith("Bearer ")) {
+            throw new AuthException("유효하지 않은 토큰입니다.");
+        }
+
         Claims claims = jwtUtil.extractClaims(bearerToken.substring(7));
+        if (claims == null) {
+            throw new AuthException("토큰이 유효하지 않습니다.");
+        }
+
         long userId = Long.parseLong(claims.getSubject());
         managerService.deleteManager(userId, todoId, managerId);
     }
